+++ node/bin/index.js
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+const run = require('../src/env0-deploy-cli');
+
+run();
