--- node/index.js
@@ -1,5 +1,0 @@
-#!/usr/bin/env node
-
-const run = require('./env0-deploy-cli');
-
-run();
